final-project/requirements.txt
flask
google-cloud-translate==2.3.1
